Fixed Editor project to use CoordinatorSet with multiply Coordinators and Global
