+++ development/Common/Utility/include/Components/CoordinatorSet.h
@@ -24,22 +24,14 @@
 //! \file
 #pragma once
 
-
-//#include "ComponentTypes.h"
 #include "Components/Coordinator.h"
-#include "Meta/Hana.h"
-#include "Metrics/Concurrency.h"
 #include <functional>
 
 
 namespace yaget::comp
 {
-    namespace internal
+    namespace internalc
     {
-        template<typename T>
-        concept is_global = requires { typename T::Policy::Global; };
-
-
         template <std::size_t TupleIndex, std::size_t MaxTupleSize, typename... Tuple>
         constexpr auto coordinator_row_combine()
         {
@@ -57,18 +49,147 @@ namespace yaget::comp
             }
         }
 
-    }
+        template<typename TupleA, typename TupleB, std::size_t TupleIndex, std::size_t MaxTupleSize>
+        constexpr auto tuple_get_union()
+        {
+            if constexpr (TupleIndex < MaxTupleSize)
+            {
+                using CurrentResult = std::tuple_element_t<TupleIndex, TupleA>;
+
+                if constexpr (yaget::meta::tuple_is_element_v<CurrentResult, TupleB>)
+                {
+                    constexpr auto result = std::tuple<CurrentResult>{};
+                    constexpr auto nextResult = tuple_get_union<TupleA, TupleB, TupleIndex + 1, MaxTupleSize>();
+                    return std::tuple_cat(result, nextResult);
+                }
+                else
+                {
+                    constexpr auto nextResult = tuple_get_union<TupleA, TupleB, TupleIndex + 1, MaxTupleSize>();
+                    return nextResult;
+                }
+            }
+            else
+            {
+                return std::tuple<>{};
+            }
+        }
+
+        // Copy tuple from Source to Target overwriting any values in Target
+        template<typename S, typename T, int N = std::tuple_size_v<std::remove_reference_t<S>>>
+        constexpr void tuple_copy(const S& source, T& target)
+        {
+            tuple_copy_if(source, target, [](auto...){ return true; });
+        }
+
+        // Copy tuple from Source to Target overwriting any values in Target
+        // if callback returns true, otherwise skip it
+        template<typename S, typename T, typename C, int N = std::tuple_size_v<std::remove_reference_t<S>>>
+        constexpr void tuple_copy_if(const S& source, T& target, C callback)
+        {
+            using ET = std::tuple_element_t<N - 1, S>;
+
+            const auto& sourceElement = std::get<ET>(source);
+            const auto& targetElement = std::get<ET>(target);
+            if (callback(sourceElement, targetElement))
+            {
+                std::get<ET>(target) = std::get<ET>(source);
+            }
+
+            if constexpr (N - 1 > 0)
+            {
+                tuple_copy_if<S, T, C, N - 1>(source, target, callback);
+            }
+        }
+
+        template<typename S, typename T>
+        constexpr void tuple_copy_if_source(const S& source, T& target)
+        {
+            tuple_copy_if(source, target, [](const auto& sourceElement, const auto& targetElement)
+            {
+                return sourceElement != nullptr;
+            });
+        }
+
+        // for_loop<std::tuple<...>>([]<std::size_t T0>()
+        template<std::size_t N, typename TCallable>
+        constexpr void for_loop(TCallable&& callable)
+        {
+            callable.template operator()<N - 1>();
+            if constexpr (N - 1 > 0)
+            {
+                for_loop<N - 1>(std::forward<TCallable>(callable));
+            }
+        }
+
+        template<typename T>
+        concept requires_global_coordinator = requires { typename T::Global; };
+
+        template <typename Coordinators, size_t Num = std::tuple_size_v<Coordinators>>
+        constexpr bool has_global_coordinator()
+        {
+            if constexpr (Num > 0)
+            {
+                using CoordinatorRow = typename std::tuple_element_t<Num - 1, Coordinators>::Policy;
+                CoordinatorRow coordinatorRow{};
+                if constexpr (requires_global_coordinator<CoordinatorRow>)
+                {
+                    return true;
+                }
+
+                return has_global_coordinator<Coordinators, Num - 1>();
+            }
+
+            return false;
+        }
+    } // namespace internalc
+
+
+    template<typename TupleA, typename TupleB>
+    using tuple_get_union_t = decltype(internalc::tuple_get_union<TupleA, TupleB, 0, std::tuple_size_v<TupleA>>());
+
+
+    namespace internalc
+    {
+        template <typename Coordinators, typename QueryRow, size_t Num = std::tuple_size_v<Coordinators>>
+        constexpr bool uses_global_coordinator()
+        {
+            if constexpr (Num - 1 > 0)
+            {
+                constexpr std::size_t coordinatorIndex = Num - 1;
+
+                using CoordinatorPolicy = typename std::tuple_element_t<coordinatorIndex, Coordinators>::Policy;
+
+                if constexpr (internalc::requires_global_coordinator<CoordinatorPolicy>)
+                {
+                    using RequestedRow = comp::tuple_get_union_t<QueryRow, typename CoordinatorPolicy::Row>;
+                    
+                    if constexpr (std::tuple_size_v<RequestedRow> > 0)
+                    {
+                        return true;
+                    }
+                }
+
+                return uses_global_coordinator<Coordinators, QueryRow, Num - 1>();
+            }
+
+            return false;
+        }
+
+    } // namespace internalc
 
     template <typename... Tuple>
     struct coordinator_row_combine
     {
-        using type = decltype(internal::coordinator_row_combine<0, std::tuple_size_v<std::remove_reference_t<std::tuple<Tuple...>>>, Tuple...>());
+        using type = decltype(internalc::coordinator_row_combine<0, std::tuple_size_v<std::remove_reference_t<std::tuple<Tuple...>>>, Tuple...>());
     };
 
     template<typename... Tuple>
     using coordinator_row_combine_t = typename coordinator_row_combine<Tuple...>::type;
 
-
+    //template<char ...C>
+    //requires (sizeof...(C)%2 == 0)
+    //constexpr std::string
+    //operator""_hex()
     template <typename... T>
     class CoordinatorSet
     {
@@ -77,104 +198,114 @@ namespace yaget::comp
         static constexpr size_t NumCoordinators = std::tuple_size_v<std::remove_reference_t<Coordinators>>;
 
         using FullRow = coordinator_row_combine_t<T...>;
-        static_assert(meta::tuple_is_unique_v<FullRow>, "Duplicate element types in CoordinatorSet FullRow");
+        //static_assert(meta::tuple_is_unique_v<FullRow>, "Duplicate element types in CoordinatorSet FullRow");
         const Strings mComponentNames = comp::db::GetPolicyRowNames<FullRow>();
 
-        template <typename R>
-        std::size_t ForEach(std::function<bool(Id_t id, R components)> callback)
+        // find all rows which contain QueryRow, and call callback for each one
+        // return True to keep iterating, otherwise return False to stop
+        template <typename QueryRow>
+        using RowCallback = std::function<bool(Id_t, QueryRow)>;
+
+        template <typename QueryRow>
+        std::size_t ForEach(RowCallback<QueryRow> callback) const
         {
-            namespace hana = boost::hana;
-            using RequestRow = R;
+            constexpr bool usesGlobal = internalc::uses_global_coordinator<Coordinators, QueryRow>();
 
-            RequestRow templateRow{};
-            using Rows = std::unordered_map<Id_t, RequestRow>;
-            Rows rows;
+            // we want to compare QueryRow against each Coordinator::FullRow,
+            // and collect same elements into one tuple type and return that.
+            // with this type we can enumerate over all rows using returned tuple.
 
-            // walk over each coordinator and extract which RequestRow component belong to coordinator
-            // and construct RowPolicy to be ingested by coordinator template args
-            meta::for_each(mCoordinators, [&rows, &templateRow]<typename C>(C& coordinator)
+            using Items = std::map<comp::Id_t, QueryRow>;
+            Items collectedItems;
+            Items collectedGlobalItem;
+
+            //---------------------------------------------------------------------------------------------
+            // collect global row if coordinator exists
+            if constexpr (usesGlobal)
             {
-                using CoordType = C;
-                using CoordinatorRow = decltype(hana::to_tuple(typename CoordType::FullRow{}));
-
-                // filter user request row by which coordinator posses that component. There is a guaranty of no duplicate types between coordinators
-                constexpr auto transformedRow = hana::filter(hana::to_tuple(RequestRow{}), [&]<typename RR>(const RR&)
+                internalc::for_loop<NumCoordinators>([this, &collectedGlobalItem]<std::size_t T0>()
                 {
-                    constexpr auto hasComponent = hana::contains(CoordinatorRow{}, RR{}) == 1;
-                    if constexpr (hasComponent)
+                    constexpr std::size_t coordinatorIndex = T0;
+
+                    using CoordinatorPolicy = typename std::tuple_element_t<coordinatorIndex, Coordinators>::Policy;
+
+                    if constexpr (internalc::requires_global_coordinator<CoordinatorPolicy>)
                     {
-                        return hana::bool_c<true>;
-                    }
-                    else
-                    {
-                        return hana::bool_c<false>;
+                        using RequestedRow = tuple_get_union_t<QueryRow, typename CoordinatorPolicy::Row>;
+                        using RequestedRowPolicy = comp::GlobalRowPolicy<RequestedRow>;
+                        
+                        if constexpr (std::tuple_size_v<RequestedRow> > 0)
+                        {
+                            auto& coordinator = GetCoordinator<coordinatorIndex>();
+                            const std::size_t numItems = coordinator.template ForEach<RequestedRowPolicy>([&collectedGlobalItem](comp::Id_t id, const auto& row)
+                            {
+                                internalc::tuple_copy(row, collectedGlobalItem[id]);
+                                return true;
+                            });
+                        }
                     }
                 });
+            }
 
-                // simple optimization of not doing anything for this coordinator
-                // it no RowPolicy
-                if constexpr (hana::size(transformedRow))
+            //---------------------------------------------------------------------------------------------
+            // we need to iterate over each coordinator and collect results
+            internalc::for_loop<NumCoordinators>([this, &collectedItems]<std::size_t T0>()
+            {
+                constexpr std::size_t coordinatorIndex = T0;
+
+                using CoordinatorPolicy = typename std::tuple_element_t<coordinatorIndex, Coordinators>::Policy;
+                if constexpr (!internalc::requires_global_coordinator<CoordinatorPolicy>)
                 {
-                    //const auto& message = fmt::format("Update Entity Id: {}", id);
-                    metrics::Channel systemChannel("Collecting Entities", YAGET_METRICS_CHANNEL_FILE_LINE);
+                    using RequestedRow = tuple_get_union_t<QueryRow, typename CoordinatorPolicy::Row>;
 
-                    // create our RowPolicy object and get type
-                    constexpr auto qrow = boost::hana::unpack(transformedRow, []<typename... T0>([[maybe_unused]] T0... args)
+                    if constexpr (std::tuple_size_v<RequestedRow> > 0)
                     {
-                        using RP = comp::RowPolicy<T0...>;
-                        return RP{};
-                    });
-                    using QueryRow = decltype(qrow);
+                        using RequestedRowPolicy = comp::RowPolicy<RequestedRow>;
 
-                    [[maybe_unused]] std::size_t numItems = coordinator.template ForEach<QueryRow>([&rows, &templateRow]([[maybe_unused]] comp::Id_t id, const auto& row)
-                    {
-                        //metrics::Channel systemChannel("ForEach Entities", YAGET_METRICS_CHANNEL_FILE_LINE);
-
-                        if constexpr (internal::is_global<CoordType>)
+                        auto& coordinator = GetCoordinator<coordinatorIndex>();
+                        const std::size_t numItems = coordinator.template ForEach<RequestedRowPolicy>([&collectedItems](comp::Id_t id, const auto& row)
                         {
-                            meta::tuple_copy(row, templateRow);
-                        }
-                        else
-                        {
-                            // get ref to the row we need to modify and copy from incoming row in to it
-                            // we don't worry about stomping already set fields, since ForEach call is per item (unique id)
-                            // and only the one for this specific coordinator (there is no overlap between coordinators per types)
-                            RequestRow& currentRow = rows[id];
-                            meta::tuple_copy(row, currentRow);
-                        }
-
-                        return true;
-                    });
+                            internalc::tuple_copy(row, collectedItems[id]);
+                            return true;
+                        });
+                    }
                 }
             });
 
-            metrics::Channel systemChannel("Ticking Entities", YAGET_METRICS_CHANNEL_FILE_LINE);
+            //---------------------------------------------------------------------------------------------
+            // iterate over collected rows and trigger callback
+            std::size_t numProcessedElements = 0;
 
-            const bool validTemplateRow = templateRow != RequestRow{};
-            if (rows.empty() && validTemplateRow)
+            if constexpr (usesGlobal)
             {
-                rows[comp::GLOBAL_ID_MARKER] = templateRow;
-            }
-
-            // Now,call call back for each element in rows
-            int itemCounter = 0;
-            for (const auto& [id, row] : rows)
-            {
-                ++itemCounter;
-
-                RequestRow requestRow = row;
-                if (validTemplateRow)
+                // only if there is no regular items (collectedItems) and we have ONE global item (collectedGlobalItem)
+                if (collectedItems.empty() && !collectedGlobalItem.empty() && collectedGlobalItem.size() == 1)
                 {
-                    meta::tuple_copy(templateRow, requestRow);
-                }
-
-                if (!callback(id, requestRow))
-                {
-                    return itemCounter;
+                    collectedItems = collectedGlobalItem;
                 }
             }
 
-            return itemCounter;
+            for (auto& [id, element] : collectedItems)
+            {
+                auto elementRow = element;
+                if constexpr (usesGlobal)
+                {
+                    const auto& globalElement = collectedGlobalItem.begin();
+                    // if we do have global elements copy them into parameter before triggering callback
+                    internalc::tuple_copy_if_source(globalElement->second, elementRow);
+                }
+
+                if (callback(id, elementRow))
+                {
+                    ++numProcessedElements;
+                }
+                else
+                {
+                    break;
+                }
+            }
+
+            return numProcessedElements;
         }
 
         template <typename C>
@@ -199,12 +330,6 @@ namespace yaget::comp
         const auto& GetCoordinator() const
         {
             return std::get<Index>(mCoordinators);
-        }
-
-        template <typename TCoordinator>
-        constexpr std::size_t GetCoordinatorIndex() const
-        {
-            return meta::Index<TCoordinator, Coordinators>::value;
         }
 
     private:
