+++ development/Tools/YagetEditor/source/EditorSystem.h
@@ -21,10 +21,10 @@
 
 namespace yaget::editor
 {
-    class EditorSystem : public comp::gs::GameSystem<comp::gs::NoEndMarker, Messaging, EmptyComponent*, BlankComponent*/*, EditorComponent* */>
+    class EditorSystem : public comp::gs::GameSystem<EditorGameCoordinatorSet, comp::gs::NoEndMarker, Messaging, EmptyComponent*, BlankComponent*/*, EditorComponent* */>
     {
     public:
-        EditorSystem(Messaging& messaging, Application& app);
+        EditorSystem(Messaging& messaging, Application& app, EditorGameCoordinatorSet& coordinatorSet);
 
     private:
         void OnUpdate(comp::Id_t id, const time::GameClock& gameClock, metrics::Channel& channel, EmptyComponent* emptyComponent, const BlankComponent* blankComponent/*, EditorComponent* editorComponent*/);
